+++ vscode/webviews/Components/EnhancedContextSettings.tsx
@@ -15,7 +15,10 @@ import { PopupFrame } from '../Popups/Popup'
 import popupStyles from '../Popups/Popup.module.css'
 import styles from './EnhancedContextSettings.module.css'
 
-interface EnhancedContextSettingsProps {}
+interface EnhancedContextSettingsProps {
+    isOpen: boolean
+    setOpen: (open: boolean) => void
+}
 
 export function defaultEnhancedContextContext(): EnhancedContextContextT {
     return {
@@ -51,11 +54,23 @@ export function useEnhancedContextEventHandlers(): EnhancedContextEventHandlersT
     return React.useContext(EnhancedContextEventHandlers)
 }
 
-const ContextGroupComponent: React.FunctionComponent<{ group: ContextGroup }> = ({ group }): React.ReactNode => {
+const ContextGroupComponent: React.FunctionComponent<{ group: ContextGroup; allGroups: ContextGroup[] }> = ({
+    group,
+    allGroups,
+}): React.ReactNode => {
+    // if there's a single group, we want the group name's basename
+    let groupName
+    if (allGroups.length === 1) {
+        const matches = group.name.match(/.+[/\\](.+?)$/)
+        groupName = matches ? matches[1] : group.name
+    } else {
+        groupName = group.name
+    }
+
     return (
         <>
-            <dt>
-                <i className="codicon codicon-folder" /> {group.name}
+            <dt title={group.name} className={styles.lineBreakAll}>
+                <i className="codicon codicon-folder" /> {groupName}
             </dt>
             <dd>
                 <ol className={styles.providersList}>
@@ -87,7 +102,7 @@ const EmbeddingsConsentComponent: React.FunctionComponent<{ provider: LocalEmbed
         <div>
             <p className={styles.providerExplanatoryText}>
                 The repository&apos;s contents will be uploaded to OpenAI&apos;s Embeddings API and then stored locally.
-                To exclude files, set up a <a href="about:blank#TODO">Cody ignore file.</a>
+                {/* To exclude files, set up a <a href="about:blank#TODO">Cody ignore file.</a> */}
             </p>
             <p>
                 <VSCodeButton onClick={onClick}>Enable Embeddings</VSCodeButton>
@@ -101,9 +116,13 @@ function contextProviderState(provider: ContextProvider): React.ReactNode {
         case 'indeterminate':
         case 'ready':
             if (provider.kind === 'embeddings' && provider.type === 'remote') {
-                return <p className={styles.providerExplanatoryText}>Inherited {provider.remoteName}</p>
+                return (
+                    <p className={classNames(styles.providerExplanatoryText, styles.lineBreakAll)}>
+                        Inherited {provider.remoteName}
+                    </p>
+                )
             }
-            return <></>
+            return <span className={styles.providerInlineState}>&mdash; Indexed</span>
         case 'indexing':
             return <span className={styles.providerInlineState}>&mdash; Indexing&hellip;</span>
         case 'unconsented':
@@ -111,7 +130,8 @@ function contextProviderState(provider: ContextProvider): React.ReactNode {
         case 'no-match':
             return (
                 <p className={styles.providerExplanatoryText}>
-                    No repository matching {provider.remoteName} on <a href="about:blank#TODO">{provider.origin}</a>
+                    {/* No repository matching {provider.remoteName} on <a href="about:blank#TODO">{provider.origin}</a> */}
+                    No repository matching {provider.remoteName} on {provider.origin}
                 </p>
             )
         default:
@@ -130,7 +150,7 @@ const ContextProviderComponent: React.FunctionComponent<{ provider: ContextProvi
             stateIcon = <i className="codicon codicon-circle-outline" />
             break
         case 'ready':
-            stateIcon = <i className="codicon codicon-check" />
+            stateIcon = <i className="codicon codicon-database" />
             break
         case 'no-match':
             stateIcon = <i className="codicon codicon-circle-slash" />
@@ -149,11 +169,13 @@ const ContextProviderComponent: React.FunctionComponent<{ provider: ContextProvi
     )
 }
 
-export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSettingsProps> = (): React.ReactNode => {
+export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSettingsProps> = ({
+    isOpen,
+    setOpen,
+}): React.ReactNode => {
     const events = useEnhancedContextEventHandlers()
     const context = useEnhancedContextContext()
     const [enabled, setEnabled] = React.useState<boolean>(useEnhancedContextEnabled())
-    const [isOpen, setOpen] = React.useState(false)
     const enabledChanged = React.useCallback(
         (event: any): void => {
             const shouldEnable = !!event.target?.checked
@@ -164,11 +186,18 @@ export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSet
         },
         [events, enabled]
     )
+
+    const hasOpenedBeforeKey = 'enhanced-context-settings.has-opened-before'
+    const hasOpenedBefore = localStorage.getItem(hasOpenedBeforeKey) === 'true'
+    if (isOpen && !hasOpenedBefore) {
+        localStorage.setItem(hasOpenedBeforeKey, 'true')
+    }
+
     return (
         <div className={classNames(popupStyles.popupHost)}>
             <PopupFrame
                 isOpen={isOpen}
-                onDismiss={() => setOpen(!isOpen)}
+                onDismiss={() => setOpen(false)}
                 classNames={[popupStyles.popupTrail, styles.enhancedContextSettingsPopup]}
             >
                 <div className={styles.enhancedContextInnerContainer}>
@@ -180,12 +209,12 @@ export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSet
                             <h1>Enhanced Context ✨</h1>
                         </label>
                         <p>
-                            Automatically include additional context about your code.{' '}
-                            <a href="about:blank#TODO">Learn more</a>
+                            Include additional code context with your message.{' '}
+                            {/* <a href="about:blank#TODO">Learn more</a> */}
                         </p>
                         <dl className={styles.foldersList}>
                             {context.groups.map(group => (
-                                <ContextGroupComponent key={group.name} group={group} />
+                                <ContextGroupComponent key={group.name} group={group} allGroups={context.groups} />
                             ))}
                         </dl>
                     </div>
@@ -198,7 +227,8 @@ export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSet
                 onClick={() => setOpen(!isOpen)}
                 title="Configure Enhanced Context"
             >
-                <i className="codicon codicon-settings" />
+                <i className="codicon codicon-sparkle" />
+                {isOpen || hasOpenedBefore ? null : <div className={styles.glowyDot} />}
             </VSCodeButton>
         </div>
     )